start test refactoring
